pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deepfake-detector"
version = "0.1.0"
description = "Clean starter scaffold for deepfake video detection"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.24",
  "pandas>=2.0",
  "opencv-python>=4.8",
  "torch>=2.1",
  "torchvision>=0.16",
  "tensorflow>=2.14",
  "face-recognition>=1.3",
  "Pillow>=10.0",
  "matplotlib>=3.7",
  "PyYAML>=6.0",
  "tqdm>=4.66",
  "scikit-learn>=1.3",
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
  "python-multipart>=0.0.9",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
